Decompose consumer into traits
