+++ modules/core/src/main/scala/fs2/kafka/KafkaConsumer.scala
@@ -19,6 +19,7 @@ import fs2.kafka.internal.converters.collection._
 import fs2.kafka.internal.instances._
 import fs2.kafka.internal.KafkaConsumerActor._
 import fs2.kafka.internal.syntax._
+import fs2.kafka.consumer._
 import java.util
 
 import org.apache.kafka.common.{Metric, MetricName, PartitionInfo, TopicPartition}
@@ -61,297 +62,14 @@ import scala.util.matching.Regex
   * per [[KafkaConsumer]], and if you want to start a new stream if the first
   * one finishes, let the [[KafkaConsumer]] shutdown and create a new one.
   */
-sealed abstract class KafkaConsumer[F[_], K, V] {
-
-  /**
-    * Alias for `partitionedStream.parJoinUnbounded`.
-    * See [[partitionedStream]] for more information.
-    *
-    * @note you have to first use `subscribe` to subscribe the consumer
-    *       before using this `Stream`. If you forgot to subscribe, there
-    *       will be a [[NotSubscribedException]] raised in the `Stream`.
-    */
-  def stream: Stream[F, CommittableConsumerRecord[F, K, V]]
-
-  /**
-    * `Stream` where the elements themselves are `Stream`s which continually
-    * request records for a single partition. These `Stream`s will have to be
-    * processed in parallel, using `parJoin` or `parJoinUnbounded`. Note that
-    * when using `parJoin(n)` and `n` is smaller than the number of currently
-    * assigned partitions, then there will be assigned partitions which won't
-    * be processed. For that reason, prefer `parJoinUnbounded` and the actual
-    * limit will be the number of assigned partitions.<br>
-    * <br>
-    * If you do not want to process all partitions in parallel, then you
-    * can use [[stream]] instead, where records for all partitions are in
-    * a single `Stream`.
-    *
-    * @note you have to first use `subscribe` to subscribe the consumer
-    *       before using this `Stream`. If you forgot to subscribe, there
-    *       will be a [[NotSubscribedException]] raised in the `Stream`.
-    */
-  def partitionedStream: Stream[F, Stream[F, CommittableConsumerRecord[F, K, V]]]
-
-  /**
-    * `Stream` where each element contains a current assignment. The current
-    * assignment is the `Map`, where keys is a `TopicPartition`, and values are
-    * streams with records for a particular `TopicPartition`.<br>
-    * <br>
-    * New assignments will be received on each rebalance. On rebalance,
-    * Kafka revoke all previously assigned partitions, and after that assigned
-    * new partitions all at once. `partitionsMapStream` reflects this process
-    * in a streaming manner.<br>
-    * <br>
-    * Note, that partition streams for revoked partitions will
-    * be closed after the new assignment comes.<br>
-    * <br>
-    * This is the most generic `Stream` method. If you don't need such control,
-    * consider using `partitionedStream` or `stream` methods.
-    * They are both based on a `partitionsMapStream`.
-    *
-    * @note you have to first use `subscribe` to subscribe the consumer
-    *       before using this `Stream`. If you forgot to subscribe, there
-    *       will be a [[NotSubscribedException]] raised in the `Stream`.
-    * @see [[stream]]
-    * @see [[partitionedStream]]
-    */
-  def partitionsMapStream
-    : Stream[F, Map[TopicPartition, Stream[F, CommittableConsumerRecord[F, K, V]]]]
-
-  /**
-    * Returns the set of partitions currently assigned to this consumer.
-    */
-  def assignment: F[SortedSet[TopicPartition]]
-
-  /**
-    * `Stream` where the elements are the set of `TopicPartition`s currently
-    * assigned to this consumer. The stream emits whenever a rebalance changes
-    * partition assignments.
-    */
-  def assignmentStream: Stream[F, SortedSet[TopicPartition]]
-
-  /**
-    * Overrides the fetch offsets that the consumer will use when reading the
-    * next record. If this API is invoked for the same partition more than once,
-    * the latest offset will be used. Note that you may lose data if this API is
-    * arbitrarily used in the middle of consumption to reset the fetch offsets.
-    */
-  def seek(partition: TopicPartition, offset: Long): F[Unit]
-
-  /**
-    * Seeks to the first offset for each currently assigned partition.
-    * This is equivalent to using `seekToBeginning` with an empty set
-    * of partitions.<br>
-    * <br>
-    * Note that this seek evaluates lazily, and only on the next call
-    * to `poll` or `position`.
-    */
-  def seekToBeginning: F[Unit]
-
-  /**
-    * Seeks to the first offset for each of the specified partitions.
-    * If no partitions are provided, seeks to the first offset for
-    * all currently assigned partitions.<br>
-    * <br>
-    * Note that this seek evaluates lazily, and only on the next call
-    * to `poll` or `position`.
-    */
-  def seekToBeginning[G[_]](partitions: G[TopicPartition])(
-    implicit G: Foldable[G]
-  ): F[Unit]
-
-  /**
-    * Seeks to the last offset for each currently assigned partition.
-    * This is equivalent to using `seekToEnd` with an empty set of
-    * partitions.<br>
-    * <br>
-    * Note that this seek evaluates lazily, and only on the next call
-    * to `poll` or `position`.
-    */
-  def seekToEnd: F[Unit]
-
-  /**
-    * Seeks to the last offset for each of the specified partitions.
-    * If no partitions are provided, seeks to the last offset for
-    * all currently assigned partitions.<br>
-    * <br>
-    * Note that this seek evaluates lazily, and only on the next call
-    * to `poll` or `position`.
-    */
-  def seekToEnd[G[_]](partitions: G[TopicPartition])(
-    implicit G: Foldable[G]
-  ): F[Unit]
-
-  /**
-    * Returns the offset of the next record that will be fetched.<br>
-    * <br>
-    * Timeout is determined by `default.api.timeout.ms`, which
-    * is set using [[ConsumerSettings#withDefaultApiTimeout]].
-    */
-  def position(partition: TopicPartition): F[Long]
-
-  /**
-    * Returns the offset of the next record that will be fetched.
-    */
-  def position(partition: TopicPartition, timeout: FiniteDuration): F[Long]
-
-  /**
-    * Subscribes the consumer to the specified topics. Note that you have to
-    * use one of the `subscribe` functions to subscribe to one or more topics
-    * before using any of the provided `Stream`s, or a [[NotSubscribedException]]
-    * will be raised in the `Stream`s.
-    */
-  def subscribeTo(firstTopic: String, remainingTopics: String*): F[Unit]
-
-  /**
-    * Subscribes the consumer to the specified topics. Note that you have to
-    * use one of the `subscribe` functions to subscribe to one or more topics
-    * before using any of the provided `Stream`s, or a [[NotSubscribedException]]
-    * will be raised in the `Stream`s.
-    *
-    * @param topics the topics to which the consumer should subscribe
-    */
-  def subscribe[G[_]](topics: G[String])(implicit G: Reducible[G]): F[Unit]
-
-  /**
-    * Subscribes the consumer to the topics matching the specified `Regex`.
-    * Note that you have to use one of the `subscribe` functions before you
-    * can use any of the provided `Stream`s, or a [[NotSubscribedException]]
-    * will be raised in the `Stream`s.
-    *
-    * @param regex the regex to which matching topics should be subscribed
-    */
-  def subscribe(regex: Regex): F[Unit]
-
-  /**
-    * Unsubscribes the consumer from all topics and partitions assigned
-    * by `subscribe` or `assign`.
-    */
-  def unsubscribe: F[Unit]
-
-  /**
-    * Manually assigns the specified list of topic partitions to the consumer.
-    * This function does not allow for incremental assignment and will replace
-    * the previous assignment (if there is one).
-    *
-    * Manual topic assignment through this method does not use the consumer's
-    * group management functionality. As such, there will be no rebalance
-    * operation triggered when group membership or cluster and topic metadata
-    * change. Note that it is not possible to use both manual partition
-    * assignment with `assign` and group assigment with `subscribe`.
-    *
-    * If auto-commit is enabled, an async commit (based on the old assignment)
-    * will be triggered before the new assignment replaces the old one.
-    *
-    * To unassign all partitions, use [[KafkaConsumer#unsubscribe]].
-    *
-    * @see org.apache.kafka.clients.consumer.KafkaConsumer#assign
-    */
-  def assign(partitions: NonEmptySet[TopicPartition]): F[Unit]
-
-  /**
-    * Manually assigns the specified list of partitions for the specified topic
-    * to the consumer. This function does not allow for incremental assignment
-    * and will replace the previous assignment (if there is one).
-    *
-    * Manual topic assignment through this method does not use the consumer's
-    * group management functionality. As such, there will be no rebalance
-    * operation triggered when group membership or cluster and topic metadata
-    * change. Note that it is not possible to use both manual partition
-    * assignment with `assign` and group assignment with `subscribe`.
-    *
-    * If auto-commit is enabled, an async commit (based on the old assignment)
-    * will be triggered before the new assignment replaces the old one.
-    *
-    * To unassign all partitions, use [[KafkaConsumer#unsubscribe]].
-    *
-    * @see org.apache.kafka.clients.consumer.KafkaConsumer#assign
-    */
-  def assign(topic: String, partitions: NonEmptySet[Int]): F[Unit]
-
-  /**
-    * Manually assigns all partitions for the specified topic to the consumer.
-    */
-  def assign(topic: String): F[Unit]
-
-  /**
-    * Returns the partitions for the specified topic.
-    *
-    * Timeout is determined by `default.api.timeout.ms`, which
-    * is set using [[ConsumerSettings#withDefaultApiTimeout]].
-    */
-  def partitionsFor(topic: String): F[List[PartitionInfo]]
-
-  /**
-    * Returns the partitions for the specified topic.
-    */
-  def partitionsFor(topic: String, timeout: FiniteDuration): F[List[PartitionInfo]]
-
-  /**
-    * Returns the first offset for the specified partitions.<br>
-    * <br>
-    * Timeout is determined by `default.api.timeout.ms`, which
-    * is set using [[ConsumerSettings#withDefaultApiTimeout]].
-    */
-  def beginningOffsets(
-    partitions: Set[TopicPartition]
-  ): F[Map[TopicPartition, Long]]
-
-  /**
-    * Returns the first offset for the specified partitions.
-    */
-  def beginningOffsets(
-    partitions: Set[TopicPartition],
-    timeout: FiniteDuration
-  ): F[Map[TopicPartition, Long]]
-
-  /**
-    * Returns the last offset for the specified partitions.<br>
-    * <br>
-    * Timeout is determined by `request.timeout.ms`, which
-    * is set using [[ConsumerSettings#withRequestTimeout]].
-    */
-  def endOffsets(
-    partitions: Set[TopicPartition]
-  ): F[Map[TopicPartition, Long]]
-
-  /**
-    * Returns the last offset for the specified partitions.
-    */
-  def endOffsets(
-    partitions: Set[TopicPartition],
-    timeout: FiniteDuration
-  ): F[Map[TopicPartition, Long]]
-
-  /**
-    * A `Fiber` that can be used to cancel the underlying consumer, or
-    * wait for it to complete. If you're using [[stream]], or any other
-    * provided stream in [[KafkaConsumer]], these will be automatically
-    * interrupted when the underlying consumer has been cancelled or
-    * when it finishes with an exception.<br>
-    * <br>
-    * Whenever `cancel` is invoked, an attempt will be made to stop the
-    * underlying consumer. The `cancel` operation will not wait for the
-    * consumer to shutdown. If you also want to wait for the shutdown
-    * to complete, you can use `join`. Note that `join` is guaranteed
-    * to complete after consumer shutdown, even when the consumer is
-    * cancelled with `cancel`.<br>
-    * <br>
-    * This `Fiber` instance is usually only required if the consumer
-    * needs to be cancelled due to some external condition, or when an
-    * external process needs to be cancelled whenever the consumer has
-    * shut down. Most of the time, when you're only using the streams
-    * provided by [[KafkaConsumer]], there is no need to use this.
-    */
-  def fiber: Fiber[F, Unit]
-
-  /**
-    * Returns consumer metrics.
-    *
-    * @see org.apache.kafka.clients.consumer.KafkaConsumer#metrics
-    */
-  def metrics: F[Map[MetricName, Metric]]
-}
+sealed abstract class KafkaConsumer[F[_], K, V]
+    extends Consume[F, K, V]
+    with Assignment[F]
+    with Offsets[F]
+    with Subscription[F]
+    with Topics[F]
+    with Metrics[F]
+    with ConsumerFiber[F]
 
 private[kafka] object KafkaConsumer {
   private[this] def startConsumerActor[F[_], K, V](
