+++ modules/core/src/main/scala/fs2/kafka/consumer/Metrics.scala
@@ -1,0 +1,13 @@
+package fs2.kafka.consumer
+
+import org.apache.kafka.common.{MetricName, Metric}
+
+trait Metrics[F[_]] {
+
+  /**
+    * Returns consumer metrics.
+    *
+    * @see org.apache.kafka.clients.consumer.KafkaConsumer#metrics
+    */
+  def metrics: F[Map[MetricName, Metric]]
+}
