+++ modules/core/src/main/scala/fs2/kafka/consumer/Offsets.scala
@@ -1,0 +1,80 @@
+/*
+ * Copyright 2018-2020 OVO Energy Limited
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package fs2.kafka.consumer
+
+import org.apache.kafka.common.TopicPartition
+import cats.Foldable
+import scala.concurrent.duration.FiniteDuration
+
+trait Offsets[F[_]] {
+
+  /**
+    * Overrides the fetch offsets that the consumer will use when reading the
+    * next record. If this API is invoked for the same partition more than once,
+    * the latest offset will be used. Note that you may lose data if this API is
+    * arbitrarily used in the middle of consumption to reset the fetch offsets.
+    */
+  def seek(partition: TopicPartition, offset: Long): F[Unit]
+
+  /**
+    * Seeks to the first offset for each currently assigned partition.
+    * This is equivalent to using `seekToBeginning` with an empty set
+    * of partitions.<br>
+    * <br>
+    * Note that this seek evaluates lazily, and only on the next call
+    * to `poll` or `position`.
+    */
+  def seekToBeginning: F[Unit]
+
+  /**
+    * Seeks to the first offset for each of the specified partitions.
+    * If no partitions are provided, seeks to the first offset for
+    * all currently assigned partitions.<br>
+    * <br>
+    * Note that this seek evaluates lazily, and only on the next call
+    * to `poll` or `position`.
+    */
+  def seekToBeginning[G[_]](partitions: G[TopicPartition])(
+    implicit G: Foldable[G]
+  ): F[Unit]
+
+  /**
+    * Seeks to the last offset for each currently assigned partition.
+    * This is equivalent to using `seekToEnd` with an empty set of
+    * partitions.<br>
+    * <br>
+    * Note that this seek evaluates lazily, and only on the next call
+    * to `poll` or `position`.
+    */
+  def seekToEnd: F[Unit]
+
+  /**
+    * Seeks to the last offset for each of the specified partitions.
+    * If no partitions are provided, seeks to the last offset for
+    * all currently assigned partitions.<br>
+    * <br>
+    * Note that this seek evaluates lazily, and only on the next call
+    * to `poll` or `position`.
+    */
+  def seekToEnd[G[_]](partitions: G[TopicPartition])(
+    implicit G: Foldable[G]
+  ): F[Unit]
+
+  /**
+    * Returns the offset of the next record that will be fetched.<br>
+    * <br>
+    * Timeout is determined by `default.api.timeout.ms`, which
+    * is set using [[ConsumerSettings#withDefaultApiTimeout]].
+    */
+  def position(partition: TopicPartition): F[Long]
+
+  /**
+    * Returns the offset of the next record that will be fetched.
+    */
+  def position(partition: TopicPartition, timeout: FiniteDuration): F[Long]
+
+}
